Set PICO_FLASH_SPI_CLKDIV=2 when building boot2
The assembly files contain a configurable clock divider, with a default
value of 4. The C SDK overwrites this value with 2, eg. for the rp pico:

https://github.com/raspberrypi/pico-sdk/blob/master/src/boards/include/boards/pico.h#L70

Because we didn't set that value in rp2040-boot2-rs, rust firmwares
effectively ran the flash chip at half speed.

Therefore, set the preprocessor define from build.rs, to match what the
C SDK does.